refactor: migrate issueSubscriber store to pinia
