+++ frontend/src/components/Issue/IssueActivityPanel.vue
@@ -354,7 +354,7 @@ import {
 import { IssueTemplate, IssueBuiltinFieldId } from "../../plugins";
 import { useI18n } from "vue-i18n";
 import dayjs from "dayjs";
-import { useUIStateStore } from "@/store";
+import { useUIStateStore, useIssueSubscriberStore } from "@/store";
 
 interface LocalState {
   showDeleteCommentModal: boolean;
@@ -472,9 +472,7 @@ export default defineComponent({
     });
 
     const subscriberList = computed((): IssueSubscriber[] => {
-      return store.getters["issueSubscriber/subscriberListByIssue"](
-        props.issue.id
-      );
+      return useIssueSubscriberStore().subscriberListByIssue(props.issue.id);
     });
 
     const cancelEditComment = () => {
