+++ frontend/src/components/Issue/IssueSubscriberPanel.vue
@@ -57,6 +57,7 @@ import {
 import PrincipalAvatar from "../PrincipalAvatar.vue";
 import { Issue, IssueSubscriber } from "../../types";
 import { useStore } from "vuex";
+import { useIssueSubscriberStore } from "@/store";
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
 interface LocalState {}
@@ -74,22 +75,18 @@ export default defineComponent({
   setup(props, { emit }) {
     const store = useStore();
     const state = reactive<LocalState>({});
+    const issueSubscriberStore = useIssueSubscriberStore();
 
     const currentUser = computed(() => store.getters["auth/currentUser"]());
 
     const prepareSubscriberList = () => {
-      store.dispatch(
-        "issueSubscriber/fetchSubscriberListByIssue",
-        props.issue.id
-      );
+      issueSubscriberStore.fetchSubscriberListByIssue(props.issue.id);
     };
 
     watchEffect(prepareSubscriberList);
 
     const subscriberList = computed((): IssueSubscriber[] => {
-      return store.getters["issueSubscriber/subscriberListByIssue"](
-        props.issue.id
-      );
+      return issueSubscriberStore.subscriberListByIssue(props.issue.id);
     });
 
     const isCurrentUserSubscribed = computed((): boolean => {
