+++ frontend/src/store/pinia-modules/index.ts
@@ -4,6 +4,7 @@ export * from "./bookmark";
 export * from "./command";
 export * from "./debug";
 export * from "./deployment";
+export * from "./issueSubscriber";
 export * from "./label";
 export * from "./notification";
 export * from "./oauth";
